requirements.txt
asgiref==3.5.2
config==0.5.1
Django==4.1.1
django-jazzmin==2.5.0
djangorestframework==3.13.1
djecrety==1.0.15
python-decouple==3.6
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2022.2
